pyproject.toml
[project]
name = "projeto-elt-dbt"
version = "0.1.0"
description = "ELT com Dbt-core e Snowflake"
authors = [
    {name = "Luisa Muzzi",email = "luisamuzzi29@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.5"
dependencies = [
    "dbt-core (>=1.10.15,<2.0.0)",
    "dbt-snowflake (>=1.10.3,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "snowflake-sqlalchemy (>=1.7.7,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
